+++ web/spa/src/pages/ProjectsPage.test.tsx
@@ -39,6 +39,7 @@ const mockProject = {
   claude_md: null,
   intent_md: null,
   ratchet_yaml: null,
+  required_capabilities: [],
 }
 
 describe('ProjectsPage', () => {